pipeline-stream: Add another way to find dai in case of failure
At the moment, pipeline_get_timestamp is called by host_dma_cb.
This is bad because pipeline_get_timestamp will not be able
to reach the DAI in order to query it for the dai_posn
because it will always be called during a graph walkthrough
done by the copy operation. This change will allow
pipeline_get_timestamp to work on platforms on which the
host copy operation automatically calls
pipeline_get_timestamp.

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>