audio: dai-zephyr: Allow address decrement/increment on NXP's EDMA
NXP's EDMA can automatically adjust the source and destination addresses
upon transfer completion. As such, we need to indicate how the adjustment
should be performed. In the case of playback, the source address
should be decremented, while in the case of capture, the destination
address should be the one being decremented.

Signed-off-by: Laurentiu Mihalcea <laurentiu.mihalcea@nxp.com>